move memoize and sign; add memThunk; memoize computeExtents
memThunk is a simple wrapper around memoize that, similar to reselect's
selectors, lets you define a "thunk" functions (functions that
take "no input") that represents computed state based off of internal
state

We use this to memoize computeExtents in plot.ts and delete
the manual _updateExtents calls. This improves
readability, maintainability and performance.

This introduced a bug - updating extents is sometimes overridden by
subclasses. This should be addressed
